Support PHP versions lower than 7.4
